refactor: move chat and from  model to a common file 🎨
Chat and from is common model used in various response
