+++ TelegramBot8/Model/Response/__init__.py
@@ -1,4 +1,5 @@
 from .BaseConvertors import *
+from .BaseResponse import *
 from .ForwardResponse import *
 from .GetMeResponse import *
 from .Response import *
